# docs/openapi-admin-v2.1.json
{
  "openapi": "3.1.0",
  "info": { "title": "剧院死神管理 API", "version": "2.1-admin.1", "description": "独立维护者会话；不得使用玩家Cookie代替。秘密token只在生成响应中出现一次。" },
  "servers": [{ "url": "/api/v2/admin" }],
  "security": [{ "adminCookie": [] }],
  "paths": {
    "/auth/login": { "post": { "security": [], "summary": "管理员登录", "requestBody": { "$ref": "#/components/requestBodies/AdminPassword" }, "responses": { "200": { "$ref": "#/components/responses/AdminSession" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/auth/logout": { "post": { "summary": "管理员退出", "responses": { "200": { "description": "已退出" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/me": { "get": { "summary": "当前管理员会话", "responses": { "200": { "$ref": "#/components/responses/AdminSession" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/summary": { "get": { "summary": "资源统计与最近50条管理操作", "responses": { "200": { "description": "管理概览" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/accounts": { "get": { "summary": "分页搜索账户", "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }, { "name": "query", "in": "query", "schema": { "type": "string" } }, { "name": "status", "in": "query", "schema": { "enum": ["all", "active", "disabled"] } }], "responses": { "200": { "description": "AdminPage<AdminAccountSummary>" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/accounts/{userId}": { "patch": { "summary": "修改账号名；对局/复盘中拒绝", "parameters": [{ "$ref": "#/components/parameters/UserId" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Intent" }, { "type": "object", "properties": { "username": { "type": "string", "pattern": "^[A-Za-z0-9_]{3,32}$" } }, "required": ["username"] }] } } } }, "responses": { "200": { "description": "更新后的公开Profile" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/accounts/{userId}/disable": { "post": { "$ref": "#/components/pathItems/AccountAction", "summary": "停用并撤销全部会话" } },
    "/accounts/{userId}/enable": { "post": { "$ref": "#/components/pathItems/AccountAction", "summary": "重新启用" } },
    "/accounts/{userId}/revoke-sessions": { "post": { "$ref": "#/components/pathItems/AccountAction", "summary": "撤销全部玩家会话" } },
    "/accounts/{userId}/clear-avatar": { "post": { "$ref": "#/components/pathItems/AccountAction", "summary": "清除头像并递增profileVersion" } },
    "/accounts/{userId}/reset-token": { "post": { "$ref": "#/components/pathItems/AccountAction", "summary": "签发30分钟密码重置码" } },
    "/invitations": {
      "get": { "summary": "分页查询注册邀请码与重置码", "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }, { "name": "purpose", "in": "query", "schema": { "enum": ["all", "register", "reset"] } }, { "name": "status", "in": "query", "schema": { "enum": ["all", "active", "used", "revoked", "expired"] } }], "responses": { "200": { "description": "AdminPage<AdminInvitationSummary>" }, "default": { "$ref": "#/components/responses/Error" } } },
      "post": { "summary": "生成注册邀请码", "requestBody": { "$ref": "#/components/requestBodies/InviteTtl" }, "responses": { "201": { "$ref": "#/components/responses/OneTimeSecret" }, "default": { "$ref": "#/components/responses/Error" } } }
    },
    "/invitations/{invitationId}": { "patch": { "summary": "修改活动邀请码到期时间", "parameters": [{ "$ref": "#/components/parameters/InvitationId" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Intent" }, { "type": "object", "properties": { "expiresAt": { "type": "integer" } }, "required": ["expiresAt"] }] } } } }, "responses": { "200": { "description": "更新成功" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/invitations/{invitationId}/revoke": { "post": { "$ref": "#/components/pathItems/InvitationAction", "summary": "撤销活动邀请码" } },
    "/invitations/{invitationId}/regenerate": { "post": { "summary": "撤销旧项并生成新秘密", "parameters": [{ "$ref": "#/components/parameters/InvitationId" }], "requestBody": { "$ref": "#/components/requestBodies/InviteTtl" }, "responses": { "201": { "$ref": "#/components/responses/OneTimeSecret" }, "default": { "$ref": "#/components/responses/Error" } } } },
    "/avatars/{assetId}": { "get": { "summary": "管理员授权读取账户头像", "parameters": [{ "name": "assetId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "WebP头像", "content": { "image/webp": {} } }, "default": { "$ref": "#/components/responses/Error" } } } }
  },
  "components": {
    "securitySchemes": { "adminCookie": { "type": "apiKey", "in": "cookie", "name": "td_admin_v2", "description": "HttpOnly、SameSite=Strict、固定2小时，服务重启后失效。" } },
    "parameters": {
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
      "UserId": { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
      "InvitationId": { "name": "invitationId", "in": "path", "required": true, "schema": { "type": "string" } }
    },
    "schemas": {
      "Intent": { "type": "object", "properties": { "requestId": { "type": "string", "minLength": 1, "maxLength": 80 } }, "required": ["requestId"] },
      "AdminSession": { "type": "object", "properties": { "authenticated": { "const": true }, "expiresAt": { "type": "integer" } }, "required": ["authenticated", "expiresAt"], "additionalProperties": false },
      "OneTimeSecret": { "type": "object", "properties": { "id": { "type": "string" }, "token": { "type": "string" }, "expiresAt": { "type": "integer" }, "purpose": { "enum": ["register", "reset"] } }, "required": ["id", "token", "expiresAt", "purpose"], "additionalProperties": false },
      "Error": { "type": "object", "properties": { "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } }, "required": ["code"] } }, "required": ["error"] }
    },
    "requestBodies": {
      "AdminPassword": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "password": { "type": "string", "minLength": 1, "maxLength": 256 } }, "required": ["password"], "additionalProperties": false } } } },
      "Intent": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Intent" } } } },
      "InviteTtl": { "required": true, "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Intent" }, { "type": "object", "properties": { "ttlSeconds": { "type": "integer", "minimum": 300, "maximum": 2592000 } }, "required": ["ttlSeconds"] }] } } } }
    },
    "responses": {
      "AdminSession": { "description": "独立管理员会话", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminSession" } } } },
      "OneTimeSecret": { "description": "token仅本次响应可见", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OneTimeSecret" } } } },
      "Error": { "description": "结构化错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "pathItems": {
      "AccountAction": { "parameters": [{ "$ref": "#/components/parameters/UserId" }], "requestBody": { "$ref": "#/components/requestBodies/Intent" }, "responses": { "200": { "description": "成功" }, "201": { "$ref": "#/components/responses/OneTimeSecret" }, "default": { "$ref": "#/components/responses/Error" } } },
      "InvitationAction": { "parameters": [{ "$ref": "#/components/parameters/InvitationId" }], "requestBody": { "$ref": "#/components/requestBodies/Intent" }, "responses": { "200": { "description": "成功" }, "default": { "$ref": "#/components/responses/Error" } } }
    }
  }
}